perf(frontend): split api service by domain and lazy-load demo panel
A Lighthouse mobile audit flagged ~60KB of unused JS in the monolithic
services/api.ts. Splitting it into per-domain modules (vehicleApi,
authApi, notificationsApi, pushApi, demoApi, mapApi) lets mapApi ship
only in MapView's lazy chunk instead of every route's eager bundle.

DemoControlPanel (and its demoApi dependency) is now loaded via
defineAsyncComponent instead of a static import, so its ~7KB chunk is
only fetched when demo mode is actually enabled.
