+++ frontend/src/main.ts
@@ -113,7 +113,7 @@ import App from './App.vue'
 import router from './router'
 import { useSettingsStore } from './stores/settings'
 import { useAuthStore } from './stores/auth'
-import { setUnauthorizedHandler, clearUnauthorizedState } from './services/api'
+import { setUnauthorizedHandler, clearUnauthorizedState } from './services/apiCore'
 import en from './locales/en.json'
 import nl from './locales/nl.json'
