+++ frontend/src/views/StatisticsView.vue
@@ -6,8 +6,8 @@ import { useVehicleStore } from '@/stores/vehicle'
 import { useSettingsStore } from '@/stores/settings'
 import { defaultStatsInsights, defaultStatsCharts } from '@/stores/settings'
 import type { StatsInsightId, StatsChartId } from '@/stores/settings'
-import { vehicleApi } from '@/services/api'
-import type { TelemetrySnapshot, VehicleAggregateStats } from '@/services/api'
+import { vehicleApi } from '@/services/vehicleApi'
+import type { TelemetrySnapshot, VehicleAggregateStats } from '@/services/vehicleApi'
 import { Line, Bar } from 'vue-chartjs'
 import { VueDraggable } from 'vue-draggable-plus'
 import { LMap, LTileLayer, LMarker } from '@vue-leaflet/vue-leaflet'
